scanner/requirements.txt
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
Pillow==11.0.0
numpy==2.1.3
